+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Now automatically stores documents in version-indexed files, when the file already
   exists.
+- Added extra tools related to counting the length of the generated text.
+- Added `--temperature` option to control the randomness of the generated text. This now
+  defaults to 0.0.
 
 ### Changed
 - Prompt improvements.